feat: 統計報表修正、前端 RWD 全面優化與單元測試
- 修正 StatisticsController 與 ReplenishmentSummaryDTO 型態 (Integer → Long)
- 修正統計報表 API 回傳 403 問題，DTO 構造器匹配 Long 型態
- 員工點貨頁面蔬菜區改為動態載入（從管理後台青菜類商品讀取）
- 商品管理頁面按分類分組展示，新增商品時可選擇分類
- 管理員後台側邊欄手機版改為抽屜式，點擊連結自動關閉
- 優化登入頁面 RWD，卡片風格統一，手機與平板自適應
- 分類管理、安全庫存、人資管理對話框標籤置頂，RWD 調整
- 新增後端單元測試 (InventoryService, InvSafeStockService, StatisticsService, EmployeeService)
- 新增 Controller 層測試 (InventoryControllerTest)，修正 CSRF 與 UUID 問題
- 新增 Dockerfile (後端) 與前端 Nginx 配置，支援 Render 部署
- 更新 .gitignore，排除前端依賴與建置產物

+++ backend/src/main/java/com/ysgs/dto/ReplenishmentSummaryDTO.java
@@ -1,14 +1,15 @@
 package com.ysgs.dto;
 
+import java.util.UUID;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import java.util.UUID;
 
 @Data
 @AllArgsConstructor
 public class ReplenishmentSummaryDTO {
     private UUID productId;
     private String productName;
-    private Integer totalQuantity;
-    private Integer count;
+    private Long totalQuantity;   // 改為 Long
+    private Long count;           // 改為 Long
 }
